Add CentOS support for role osquery
